fix test run errors
