+++ .github/workflows/build-test.yml
@@ -97,4 +97,4 @@ jobs:
               run: pnpm install --frozen-lockfile
 
             - name: Test
-              run: pnpm run test-ci
+              run: pnpm run test-scaffold && pnpm run test-ci
